+++ packages/amplify-cli/CHANGELOG.md
@@ -1,4 +1,4 @@
-# v2.0.0-rc3 (Jul 24, 2020)
+# v2.0.0
 
  * BREAKING CHANGE: Dropped support for Node.js 10.12.0 and older.
    ([CLI-89](https://jira.axway.com/browse/CLI-89))